Merged spike sorting script with python script
Now you only need to call rickshaw_preprocess_HR.py to perform both the preprocessing steps and the spike sorting step. Remember to have mountainlab installed in the conda environment when running this script. Also it is best to run this script from bash directly instead of IPython in Spyder