Method 4 for channel normalization
Same as method 3 for 1 x 32 channel maps. For 2 x 16, Modified by converting an "effective map" as 1x16 since if either of the electrodes is available, the spike can be picked up. In principle we don’t have to do it but this slight modification can take advantage of the 2x16 layout that is less prone to channel loss due to two electrodes at each depth